Major - function bridgeAllJsonDumpsToArcTables works correctly
The primary function of this module, bridgeAllJsonDumpsToArcTables, now works correctly for all test datasets.

+++ BridgeMongo2PostGreSQL.py
@@ -10,10 +10,7 @@ print 'started'
 import arcpy
 import os, json
 from station import Station
-from station import requiredFields as staRequiredFields
-from station import getStationsFieldList as staGetStationsFieldList
 from gage import Gage
-from gage import requiredFields as gageRequiredFields
 print 'imports complete'
 
 mxd = None
@@ -79,22 +76,20 @@ def getLayersDict(mxd):
 def getGDBfromDataSource(layer):
     delim = '\\'
 
-def ensureTableLayerHasFields(layerObj, fieldList):
-    myDataSource = layerObj
-    aType = type(layerObj)
-    useless = 'hi'
-    bType = type(useless)
-    if not(isinstance(layerObj, basestring)):
-        myDataSource = layerObj.dataSource
+def ensureTableLayerHasFields(layerObj, requiredFieldList):
+    myDataSource = layerObj.dataSource
     existingFields = arcpy.ListFields(myDataSource)
-    names = [f.name for f in existingFields]
-    for requiredField in fieldList:
+    existFNames = [f.name for f in existingFields]
+    for requiredField in requiredFieldList:
         fieldName = requiredField[0]
-        if fieldName not in names:
+        if fieldName not in existFNames:
             fieldType = requiredField[1]
             fLen = requiredField[2]
-            arcpy.AddField_management(layerObj, fieldName, fieldType,
+            try:
+                arcpy.AddField_management(layerObj, fieldName, fieldType,
                                       field_length=fLen)
+            except:
+                i = 0
 
 def initializeTables(mxd):
     lyrDict = getLayersDict(mxd)
@@ -102,13 +97,15 @@ def initializeTables(mxd):
     workingGdb = os.path.dirname(lyrStations.dataSource)
 
     try:
-        ensureTableLayerHasFields(lyrStations, staRequiredFields)
+        ensureTableLayerHasFields(lyrStations,
+                                  Station.getRequiredFieldsTuples())
     except:
         pass
 
     try:
         fullPath = os.path.join(workingGdb, tblGages)
-        ensureTableLayerHasFields(fullPath, gageRequiredFields)
+        ensureTableLayerHasFields(fullPath,
+                                  Gage.getRequiredFieldsTuples())
     except:
         pass
 
@@ -130,21 +127,39 @@ def _getFirstOrDefault(aCollection):
 
 def ensureAllColumns(layerObj, theType=''):
     if theType == '': return
+    if layerObj is None: return
     if 'STATIO' in theType.upper():
-        ensureTableLayerHasFields(layerObj, staRequiredFields)
+        ensureTableLayerHasFields(layerObj, Station.getRequiredFieldsTuples())
     elif 'GAGE' in theType.upper():
-        ensureTableLayerHasFields(layerObj, gageRequiredFields)
+        ensureTableLayerHasFields(layerObj, Gage.getRequiredFieldsTuples())
 
 def readStationsJson(stationsFileName):
+    if stationsFileName is None: return {}
     stationsDict = {}
-    with open(stationsFileName, 'r') as inFile:
-        allLines = inFile.readlines()
-        for aLine in allLines:
-            # rowDict = json.loads(aLine)
-            aStation = Station(aLine)
-            stationsDict[aStation._id] = aStation
+    try:
+        with open(stationsFileName, 'r') as inFile:
+            allLines = inFile.readlines()
+            for aLine in allLines:
+                aStation = Station(aLine)
+                stationsDict[aStation._id] = Station(aLine)
+    except:
+        return {}
 
     return stationsDict
+
+def readGagesJson(gagesFileName):
+    if gagesFileName is None: return {}
+    gagesDict = {}
+    try:
+        with open(gagesFileName, 'r') as inFile:
+            allLines = inFile.readlines()
+            for aLine in allLines:
+                aGage = Gage(aLine)
+                gagesDict[aGage._id] = aGage
+    except:
+        return {}
+
+    return gagesDict
 
 def getAllIds(layerObj):
     '''
@@ -161,9 +176,12 @@ def getAllIds(layerObj):
     return idSet
 
 def updateStations(stationsLayer, stationsDict):
-    deleteAllRows(stationsLayer)
+    if stationsLayer is None or len(stationsDict) == 0:
+        return
 
-    fieldList = staGetStationsFieldList()
+    fieldList = Station.getRequiredFieldNames()
+    ensureAllColumns(stationsLayer, 'stations')
+    deleteAllRows(stationsLayer)
 
     # write the new stations to the Layer
     with arcpy.da.InsertCursor(stationsLayer, fieldList) as cursor:
@@ -172,30 +190,80 @@ def updateStations(stationsLayer, stationsDict):
             cursor.insertRow(rowList)
     del cursor
 
-def readAllJsonDumps(stationsLayer, stationsName): #, gagesName):
-    stationsDict = readStationsJson(stationsName)
+def updateGages(gagesTable, gagesDict):
+    if gagesTable is None or len(gagesDict) == 0:
+        return
+
+    fieldList = Gage.getRequiredFieldNames()
+    ensureAllColumns(gagesTable, 'gages')
+    deleteAllRows(gagesTable)
+
+    # write the new stations to the Layer
+    with arcpy.da.InsertCursor(gagesTable, fieldList) as cursor:
+        for gageID, aGage in gagesDict.iteritems():
+            try:
+                cursor.insertRow(aGage.createOrUpdateRow())
+            except Exception as ex:
+                i = 0
+    del cursor
+
+def bridgeAllJsonDumpsToArcTables(stationsLayer, stationsJsonName,
+                     gagesTable, gagesJsonName): #, gagesName):
+    '''
+    This is the main function of the module. Call this to do what you want to do.
+    Given the parameters, bridge the data from the MongoDb Json dump file (which is the
+    input) to the ArcGIS tables behind the layers.
+    :param stationsLayer: Map Layer of the Stations feature class to populate
+    :param stationsJsonName: File name of the MongoDb Json Dump to be bridged for Station data
+    :param gagesTable: Table View of the Gages table to populate
+    :param gagesJsonName: File name of the MongoDb Json Dump to be bridged for Gage data
+    :return: None
+    '''
+    stationsDict = readStationsJson(stationsJsonName)
     updateStations(stationsLayer, stationsDict)
+    gagesDict = readGagesJson(gagesJsonName)
+    updateGages(gagesTable, gagesDict)
+
+def getTableByName(theMxd, tableName):
+    '''
+    Given a map document object and the name of a table view in that map,
+    return the tableView object.
+    :param theMxd: The Map Document Object where to find the table view.
+    :param tableName: The name (string) of the Table View.
+    :return:
+    '''
+    tbls = arcpy.mapping.ListTableViews(theMxd)
+    return _getFirstOrDefault(
+        [tbl for tbl in tbls if tbl.name == tableName])
+
 
 if __name__ == '__main__':
+
     testDir = 'Data/arcmapStuff'
     cwd = os.getcwd()
     testFullPath = os.path.join(cwd, testDir)
     mapFileName = os.path.join(testFullPath, 'testBed.mxd')
+
+    # swap the next two lines commented status to test different file sets.
     # frfDir = os.path.join(cwd, 'Data/FRFdata')
     frfDir = os.path.join(cwd, 'Data/SMS_json_v20180412')
     stationsFname = os.path.join(frfDir, 'stations.json')
     stationsJsonDump = os.path.join(testFullPath, stationsFname)
+    gagesFname = os.path.join(frfDir, 'gages.json')
+    gagesJsonDump = os.path.join(testFullPath, gagesFname)
 
     mxd = arcpy.mapping.MapDocument(mapFileName)
     stationsLayer = _getFirstOrDefault(
         [L for L in arcpy.mapping.ListLayers(mxd)
             if L.name == 'Station'])
+
+    gagesTable = getTableByName(mxd, 'tbl_gages')
+
     # deleteAllFields(stationsLayer, shouldRun=True)
-    stationFields = arcpy.ListFields(stationsLayer.dataSource)
-    if len(stationFields) <= 2:
-        ensureTableLayerHasFields(stationsLayer, staRequiredFields)
+    # ensureTableLayerHasFields(stationsLayer, staRequiredFields)
     # deleteAllRows(stationsLayer)
-    readAllJsonDumps(stationsLayer, stationsJsonDump)
+    bridgeAllJsonDumpsToArcTables(stationsLayer, stationsJsonDump,
+                     gagesTable, gagesJsonDump)
 
     # print arcpy.GetCount_management(stationsLayer), 'before deleteAllRows'
     # deleteAllRows(stationsLayer)
